refactor: 챌린지 리팩토링 진행
* 일반 컴포넌트와 페이지 레벨 컴포넌트 분리
* 불필요한 컴포넌트 정리
* 함수와 컴포넌트 분리
* 주석 추가 - props, state, args, return
* 기타 함수 개선 및 코드 압축
