+++ client/src/components/compete/comp/Timer.js
@@ -2,6 +2,7 @@ import styled from "styled-components";
 import moment from "moment";
 import Countdown from "react-countdown";
 
+//[styled comp] : 타이버 컨테이너
 const TimerBox = styled.div`
   height: 25px;
 
@@ -12,12 +13,14 @@ const TimerBox = styled.div`
   }
 `;
 
+//[function] : 타이머를 위한 시간 표기 형식 변환 함수
+//[args] : due-만료시간
 const renderer = ({ days, hours, minutes, seconds, completed }) => {
   if (completed) {
-    // Render a completed state
+    //완료시 처리할 내용
     return <>done!</>;
   } else {
-    // Render a countdown
+    //카운트 다운 출력
     return (
       <span>
         {days != 0 && days + "일 "}
@@ -29,9 +32,19 @@ const renderer = ({ days, hours, minutes, seconds, completed }) => {
   }
 };
 
+/**
+ * [comp]
+ * Timer
+ *
+ * [state]
+ * none
+ *
+ * [props]
+ * 챌린지 정보-만료일
+ */
+
 const Timer = (props) => {
   const now = moment();
-  const goal = moment();
 
   return <Countdown date={now + (props.due - now)} renderer={renderer} />;
 };
