+++ client/src/components/compete/pages/CompLists.js
@@ -1,12 +1,29 @@
 import ContentControlBtn from "components/common/ContentControlBtn";
 import React, { useState } from "react";
 import styled from "styled-components";
-// import { SetType, Type } from "./comp/SetCategory";
 import AllList from "./CompAllList";
 import MyList from "./CompMyList";
+
+//세부 옵션 리스트
 const allControlNameList = ["인기순", "최신순", "마감"];
 const myControlNameList = ["참가중", "참가완료"];
 
+//[styled comp] : 세부 카테고리 설정 옵션 박스
+const SetTypeStyle = styled.div`
+  margin: 24px 28px 0;
+`;
+
+/**
+ * [comp]
+ * CompList
+ *
+ * [state]
+ * filter : 전체 챌린지 구분용 state
+ * Done : 내 챌린지 구분용 state
+ *
+ * [props]
+ * category: 현재 선택된 카테고리 정보 - bool
+ */
 function CompList(props) {
   //리스트별 카테고리 기본 설정값
   const [filter, setFilter] = useState(allControlNameList[0]);
@@ -63,7 +80,4 @@ function CompList(props) {
   );
 }
 
-const SetTypeStyle = styled.div`
-  margin: 24px 28px 0;
-`;
 export default CompList;
