+++ client/src/components/compete/pages/Compete.js
@@ -1,0 +1,68 @@
+import { Header } from "../../common/Header";
+import AllCompContext from "../context/AllCompContext";
+import MyCompeteContext from "../context/MyCompContext";
+import NavBar from "components/common/NavBar";
+import styled from "styled-components";
+import { allCompList, myCompList } from "../context/data";
+import React, { useState } from "react";
+import CompLists from "./CompLists";
+import Banner from "../comp/BannerSwiper";
+import CategoryButton from "../comp/CategoryButton";
+
+//[styled comp] : 임시 하단 nav
+const StyledNavBar = styled(NavBar)`
+  position: fixed;
+  bottom: 0;
+`;
+
+//[styled comp]: 상위 패딩 무시용 박스
+const IgnorePaddingBox = styled.div`
+  margin: 0 -20px;
+`;
+
+/**
+ * [comp]
+ * Compete
+ *
+ * [state]
+ * isAll : 카테고리 설정을 위한 state
+ *
+ * [props]
+ * none
+ */
+function Compete() {
+  const [isAll, setisAll] = useState(true);
+
+  function setCategoryWrapper(input) {
+    setisAll(input);
+  }
+
+  return (
+    <>
+      <MyCompeteContext.Provider value={myCompList.compList}>
+        <AllCompContext.Provider value={allCompList.compList}>
+          <Header title={false} keys={3000} alarm={true}></Header>
+          <IgnorePaddingBox>
+            <CategoryButton
+              disabled={isAll}
+              onClick={() => setCategoryWrapper(true)}
+            >
+              전체 챌린지
+            </CategoryButton>
+            <CategoryButton
+              disabled={!isAll}
+              onClick={() => setCategoryWrapper(false)}
+            >
+              내 챌린지
+            </CategoryButton>
+            {isAll ? <Banner></Banner> : <></>}
+            <CompLists category={isAll}></CompLists>
+          </IgnorePaddingBox>
+        </AllCompContext.Provider>
+      </MyCompeteContext.Provider>
+      <StyledNavBar></StyledNavBar>
+    </>
+  );
+}
+
+export default Compete;
