+++ client/src/route/Routes.js
@@ -3,9 +3,9 @@ import { Routes, Route } from "react-router-dom";
 import Home from "components/home/Home";
 import Gather from "components/gather/Gather";
 import Save from "components/save/Save";
-import Compete from "components/compete/Compete";
+import Compete from "components/compete/pages/Compete";
 import Profile from "components/profile/Profile";
-import CompeteDetail from "components/compete/CompDetail";
+import CompeteDetail from "components/compete/pages/CompDetail";
 import AddMilitarySavings from "components/gather/addMilitarySavings/AddMilitarySavings";
 import AvailableSavingItemDetail from "components/gather/addMilitarySavings/AvailableSavingItemDetail";
 import AddMilitarySavingsTerm from "components/gather/addMilitarySavings/AddMilitarySavingsTerm";
@@ -15,7 +15,7 @@ import SafeBox from "components/gather/safebox/SafeBox";
 import Goal from "components/gather/addGoal/Goal";
 import Complete from "components/gather/addGoal/Complete";
 import MobileKeypad from "components/gather/safebox/MobileKeypad";
-import CompHowTo from "components/compete/CompHowTo";
+import CompHowTo from "components/compete/pages/CompHowTo";
 import KeyPopUp from "components/common/KeyPopUp";
 import Reward from "components/reward/Reward";
 import MyReward from "components/reward/MyReward";
